+++ ErsatzTV.Core/Domain/NormalizeLoudnessMode.cs
@@ -3,6 +3,5 @@ namespace ErsatzTV.Core.Domain;
 public enum NormalizeLoudnessMode
 {
     Off = 0,
-    LoudNorm = 1,
-    DynAudNorm = 2
+    LoudNorm = 1
 }